+++ src/types.ts
@@ -46,18 +46,18 @@ export type RolesAt = {
 
 export interface AccessClaims {
   name?: string;
-  user?: string;
+  user: string;
   client?: string;
   roles: RolesAt;
   scopes: Scopes[];
 }
 export interface RefreshClaims {
-  owner: string;
+  user: string;
   grant_type: GrantTypes;
 }
 export interface Claims extends AccessClaims, RefreshClaims {
-  exp: Date;
-  iat: Date;
+  exp: number; // seconds since the epoch
+  iat: number; // seconds since the epoch
 }
 
 export interface ResourceAttributeAsMapConstructor extends Array<Resource | string[]> {